Enhance FluentMenu with programmatic open/close functionality and target positioning
