--- src/Core/Components/Menu/FluentMenu.razor.ts
@@ -1,29 +1,0 @@
-export namespace Microsoft.FluentUI.Blazor.Menu {
-  export function Initialize(id: string, triggerId: string) {
-    const trigger = document.getElementById(triggerId);
-
-    if (trigger) {
-      trigger.style["anchor-name" as any] = `--anchor-${triggerId}`;
-
-      const menu = document.getElementById(id) as any;
-      if (menu && menu.slottedMenuList.length) {
-        menu.slottedMenuList[0].style["position-anchor" as any] = `--anchor-${triggerId}`;
-        menu.slottedTriggersChanged(menu.slottedTriggers, [trigger]);
-      }
-    }
-  }
-
-  export function CloseMenu(id: string) {
-    const menu = document.getElementById(id) as any;
-    if (menu) {
-      menu.closeMenu();
-    }
-   }
-
-  export function OpenMenu(id: string) {
-    const menu = document.getElementById(id) as any;
-    if (menu) {
-      menu.openMenu();
-    }
-  }
-}
